pyproject.toml
[tool.poetry]
name = "brandon"
version = "0.1.0"
description = "Generate application stubs and documentation for command line applications from YAML specifications."
authors = ["Weslley Morellato Bueno <morellato.weslley@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/wmorellato/brandon"

[tool.poetry.scripts]
brandon = "brandon.main:cli"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
black = "^22.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
